# api/adminapi/openapi.yaml
openapi: 3.0.2
info:
  title: Federation Admin
  version: 0.1.0
  description: >-
    This API enables management of federated entity data and metadata within an OpenID Federation
    environment.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8765
    description: Local development server
...
paths:
...
  /api/v1/admin/entity-configuration/lifetime:
    get:
      tags:
        - Entity Configuration
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifetimeSeconds'
          description: Successful response returning the entity-configuration lifetime in seconds.
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: getEntityConfigurationLifetime
      summary: Get entity configuration lifetime
    put:
      requestBody:
        content:
          text/plain:
            schema:
              $ref: '#/components/schemas/LifetimeSeconds'
        required: true
      tags:
        - Entity Configuration
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifetimeSeconds'
          description: Successfully updated the entity-configuration lifetime in seconds.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: updateEntityConfigurationLifetime
      summary: Update entity configuration lifetime
...
components:
  schemas:
...
    ErrorResponse:
      title: Root Type for ErrorResponse
      description: Standard error response with code and description.
      required:
        - error
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
      example:
        error: not found
        error_description: resource not found
...
    LifetimeSeconds:
      description: Lifetime value in seconds.
      type: integer
      minimum: 0
      example: 86400
...
  responses:
    BadRequestError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing parameter:
              value:
                error: invalid_request
                error_description: parameter 'foobar' is missing
      description: Invalid request parameters
...
    ServerError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server error:
              value:
                error: server_error
                error_description: foobar
      description: Internal server error
...